feat(skills): promote Skills to a dedicated center-pane tab
Skills was buried as the first row of the Settings dialog. Move it out to a
first-class 'Skills' tab in the center pane (Chat | Files | Skills), reached
right from the new-session screen. New SkillsPage is a browsable catalog:
a hero with live enabled/total/category counts, category shelves instead of
one flat list, and cards showing the skill id (mono), description, tags, a
source badge (bundled/learned/installed derived from location), and the
enable toggle — dimmed when disabled so state is scannable. Same real
endpoints (app.skills / app.skill.write / permission.skill) + the same
scratch/upload/GitHub-import flows as the old panel. Removed the Skills
settings panel + its registry entry; default settings panel is now Connectors.
